added placeholder structs for index reader and writer

+++ Project_Code_Base/John_Object_Transfer_Compiling/csharpBindings/ClientFactoryWrapperClientConfig.cs
@@ -315,7 +315,7 @@ namespace Pravega
             get
             {
                 if (i >= Count) throw new IndexOutOfRangeException();
-                var size = Marshal.SizeOf(typeof(byte));
+                var size = Marshal.SizeOf(typeof(ushort));
                 var ptr = new IntPtr(slice_pointer.ToInt64() + i * size);
                 return Marshal.PtrToStructure<byte>(ptr);
             }
